wg_service/pyproject.toml
[tool.poetry]
name = "wg_service"
version = "0.1.0"
description = ""
authors = ["avarfolomeev <avarfolomeev@sfxdx.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
setuptools = "^72.2.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
wg-wrapper = {path = "../wg_wrapper"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
